# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a comprehensive booking/reservation service API for managing property reservations with full CRUD operations, status management, and availability checking.",
        "title": "Hostflow Booking Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@hostflow.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "hostflow.software/booking",
    "basePath": "/",
    "paths": {
...
        "/reservations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all reservations for the authenticated organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Get all reservations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/booking.ReservationResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new reservation with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Create a new reservation",
                "parameters": [
                    {
                        "description": "Reservation details",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reservations/{id}": {
            "get": {
                "description": "Get reservation details by its integer ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Get reservation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update reservation details by integer ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Update a reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated reservation details",
                        "name": "reservation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "reservations"
                ],
                "summary": "Delete a reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/reservations/{id}/cancel": {
            "post": {
                "tags": [
                    "reservations"
                ],
                "summary": "Cancel a reservation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Reservation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationResponse"
                        }
                    }
                }
            }
        },
        "/reservations/{id}/status": {
            "patch": {
                "description": "Update the status of a reservation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reservations"
                ],
                "summary": "Update reservation status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reservation ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "status",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/booking.StatusUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/booking.ReservationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/booking.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "booking.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "The provided data is invalid"
                }
            }
        },
        "booking.ReservationRequest": {
            "type": "object",
            "required": [
                "check_in_date",
                "check_out_date",
                "customer_id",
                "no_of_guests",
                "organization_id",
                "property_id",
                "total_price"
            ],
            "properties": {
                "additional_requests": {
                    "type": "object",
                    "additionalProperties": true
                },
                "check_in_date": {
                    "type": "string",
                    "example": "2024-12-20T15:00:00Z"
                },
                "check_out_date": {
                    "type": "string",
                    "example": "2024-12-25T11:00:00Z"
                },
                "customer_id": {
                    "type": "integer",
                    "example": 100
                },
                "guest_data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "no_of_guests": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 2
                },
                "organization_id": {
                    "type": "integer",
                    "example": 1
                },
                "price_elements": {
                    "type": "object",
                    "additionalProperties": true
                },
                "property_id": {
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "type": "string",
                    "example": "CREATED"
                },
                "total_price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 500
                }
            }
        },
        "booking.ReservationResponse": {
            "type": "object",
            "properties": {
                "additional_requests": {
                    "type": "object",
                    "additionalProperties": true
                },
                "check_in_date": {
                    "type": "string",
                    "example": "2024-12-20T15:00:00Z"
                },
                "check_out_date": {
                    "type": "string",
                    "example": "2024-12-25T11:00:00Z"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-12-01T09:00:00Z"
                },
                "customer_id": {
                    "type": "integer",
                    "example": 100
                },
                "guest_data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "no_of_guests": {
                    "type": "integer",
                    "example": 2
                },
                "organization_id": {
                    "type": "integer",
                    "example": 1
                },
                "price_elements": {
                    "type": "object",
                    "additionalProperties": true
                },
                "property_id": {
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "CREATED CONFIRMED PAYMENT_REQUIRED REJECTED CANCELLED COMPLETED"
                    ],
                    "example": "CREATED"
                },
                "total_price": {
                    "type": "number",
                    "example": 500
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-12-01T09:00:00Z"
                }
            }
        },
        "booking.StatusUpdateRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "CREATED",
                        "CONFIRMED",
                        "PAYMENT_REQUIRED",
                        "REJECTED",
                        "CANCELLED",
                        "COMPLETED"
                    ],
                    "example": "CONFIRMED"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}